refactor: update hooks and imports to use new store structure
