+++ client/src/lib/store/hooks.ts
@@ -1,5 +1,5 @@
 import { useDispatch, useSelector, useStore } from 'react-redux'
-import type { AppDispatch, AppStore, RootState } from './store'
+import type { AppDispatch, AppStore, RootState } from '.'
 
 export const useAppDispatch = useDispatch.withTypes<AppDispatch>()
 export const useAppSelector = useSelector.withTypes<RootState>()
